+++ inspectors/snap/snap_assertion.go
@@ -68,10 +68,6 @@ func (SnapAssertionInspector) ID() string {
 
 // InspectRequest verifies if the request complies with policy.
 func (ins *SnapAssertionInspector) InspectRequest(a RequestArtifact) error {
-	if !a.RequestHeaderContains("Accept", "application/x.ubuntu.assertion") {
-		return nil
-	}
-
 	u, err := url.Parse(a.DownloadURL())
 	if err != nil {
 		return fmt.Errorf("cannot parse URL: %s", err)
@@ -97,6 +93,8 @@ func (ins *SnapAssertionInspector) InspectRequest(a RequestArtifact) error {
 		reason = "valid URL for account-key assertion download"
 	} else if _, err := newAccountKeyAssertionURLInfo(u); err == nil {
 		reason = "valid URL for account-key assertion download"
+	} else if _, err := newSerialAssertionURLInfo(u); err == nil {
+		reason = "valid URL for serial assertion download"
 	} else {
 		return nil // we don't recognize this request
 	}
@@ -150,6 +148,8 @@ func (ins *SnapAssertionInspector) InspectArtifact(f ArtifactReader, a ResponseA
 		mtype = mimetypes.AccountAssertion
 	case "account-key":
 		mtype = mimetypes.AccountKeyAssertion
+	case "serial":
+		mtype = mimetypes.SerialAssertion
 	}
 
 	a.SetArtifactMetadata(ArtifactMetadata{